% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusplus.R
\name{plot_clus_coord}
\alias{plot_clus_coord}
\title{Print the cluster coordinates plot. This graph is useful to quickly describe main characteristics of each cluster.}
\usage{
plot_clus_coord(cluster_model, data)
}
\arguments{
\item{cluster_model}{Is the cluster model developed. Only supported with k-means model for now.}

\item{data}{Is the data used to develop the model.}
}
\description{
Print the cluster coordinates plot. This graph is useful to quickly describe main characteristics of each cluster.
}

